Add files via upload
jewelry company wants to put in a bid to purchase a large set of diamonds but is unsure how much
it should bid. In this project, results from a predictive model to make a recommendation on how much
the jewelry company should bid for the diamonds.
In this project, data of the diamonds and new diamonds have on common columns, in order to
predict the new diamonds, the price we need to use diamonds, therefore we have to use common
features. In this project, Univariate analysis and Bivariate analysis have been done to find the missing
value outliers and treat them and discover correction or feature selection. As a result, price and carat
have been cleaned by deleting outliers and log transformation in order to meet the linear regression
assumptions and other models used in this project such as stepwise regression, ANN and Decision
Tree.
After applying the models, the cross-validation has been deployed to validate the result of the
initiated model, then we ensemble the models both with itself or other models. The result of
the comparison shows that the ensemble of the ANN model is the best choice for the prediction. And we
use them to predict the price of the new diamonds.